requirements.txt
asgiref==3.6.0
cssselect2==0.7.0
Django==4.1.7
django-filer==2.2.4
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
lxml==4.9.2
Pillow==9.4.0
pytz==2022.7.1
reportlab==3.6.12
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
Unidecode==1.1.2
webencodings==0.5.1
